Refactor cosmicpi-server.py into multiple files, one class per file
