+++ cosmicpi-server.py
@@ -7,247 +7,16 @@ julian.lewis lewis.julian@gmail.com 26/Feb/2016
 """
 
 import sys
-import socket
-import select
-import serial
 import time
-import traceback
 import os
-import termios
 import fcntl
-import re
-import ast
 from optparse import OptionParser
-import httplib, urllib
 
-
-# Handle keyboard input
-
-class KeyBoard(object):
-    def __init__(self):
-        self.fd = sys.stdin.fileno()
-
-    def echo_off(self):
-        self.oldterm = termios.tcgetattr(self.fd)
-        self.newattr = termios.tcgetattr(self.fd)
-        self.newattr[3] = self.newattr[3] & ~termios.ICANON & ~termios.ECHO
-        termios.tcsetattr(self.fd, termios.TCSANOW, self.newattr)
-        self.oldflags = fcntl.fcntl(self.fd, fcntl.F_GETFL)
-        fcntl.fcntl(self.fd, fcntl.F_SETFL, self.oldflags | os.O_NONBLOCK)
-
-    def echo_on(self):
-        termios.tcsetattr(self.fd, termios.TCSAFLUSH, self.oldterm)
-        fcntl.fcntl(self.fd, fcntl.F_SETFL, self.oldflags)
-
-    def test_input(self):
-        res = False
-
-        try:
-            c = sys.stdin.read(1)
-            if c == '>':
-                res = True
-        except IOError:
-            pass
-        return res
-
-
-class Socket_io(object):
-    def __init__(self, ipport):
-        try:
-            self.sik = socket.socket(socket.AF_INET, socket.SOCK_DGRAM)
-            self.sik.setsockopt(socket.SOL_SOCKET, socket.SO_REUSEPORT, 1)
-            self.sik.setblocking(0)
-            self.sik.bind(("", ipport))
-
-        except Exception, e:
-            msg = "Exception: Can't open Socket: %s" % (e)
-            print msg
-            sys.exit(1)
-
-    def recv_event_pkt(self):
-        try:
-            available = select.select([self.sik], [], [], 1)
-            if available[0]:
-                recv = self.sik.recvfrom(2048)
-                return recv
-
-        except Exception, e:
-            msg = "Exception: Can't recvfrom: %s" % (e)
-            print msg
-
-        return ["", ""]
-
-    def close(self):
-        self.sik.close()
-
-
-# Send notifications to registered mobile phones
-# Currently I am using pushover and the client has to install the
-# pushover app and register a user and application token.
-# The client supplies the keys on the cosmic pi client launch command line
-
-class Notifications(object):
-    def send_ntf(self, kyp, msg):
-        nstr = kyp.split('-')
-
-        self.conn = httplib.HTTPSConnection("api.pushover.net:443")
-
-        self.conn.request("POST", "/1/messages.json",
-                          urllib.urlencode({"token": nstr[1],
-                                            "user": nstr[0],
-                                            "sound": "Cosmic",
-                                            "message": msg}),
-                          {"Content-type": "application/x-www-form-urlencoded"})
-
-        self.conn.getresponse()
-
-
-# Each cosmic pi client can register with the server
-# We check package sequence numbers, hardware status
-
-class Registrations(object):
-    def __init__(self):
-
-        self.reg = {"Ipa": "s", "Sqn": 0, "Pat": "s", "Ntf": False, "Htu": "0", "Bmp": "0", "Acl": "0", "Mag": "0"}
-        self.regs = []
-
-    def get_len(self):
-        return len(self.regs)
-
-    def get_index_by_value(self, knam, kval):
-        if self.reg.has_key(knam):
-            for i in range(0, len(self.regs)):
-                if self.regs[i][knam] == kval:
-                    return i
-        return False
-
-    def get_reg_by_value(self, knam, kval):
-        if self.reg.has_key(knam):
-            for i in range(0, len(self.regs)):
-                if self.regs[i][knam] == kval:
-                    return self.regs[i]
-        return False
-
-    def get_reg_by_index(self, indx):
-        if indx in range(0, len(self.regs)):
-            return self.regs[indx]
-
-        return False
-
-    def get_create_reg(self, knam, kval):
-        r = self.get_reg_by_value(knam, kval)
-        if r == False:
-            i = len(self.regs)
-            self.regs.append(self.reg.copy())
-            self.regs[i][knam] = kval
-            return self.regs[i]
-        else:
-            return r
-
-    def set_reg(self, r):
-        i = self.get_index_by_value("Ipa", r["Ipa"])
-        if i == False:
-            return False
-        self.regs[i] = r.copy()
-        return True
-
-
-# This is the event object, it builds a dictionary from incomming json strings
-# and provides access to the dictionary entries containing the data for each field.
-
-class Event(object):
-    def __init__(self):
-
-        # These are the UDP packets containing json strings we are expecting
-
-        self.temperature = {"temperature": "0.0", "humidity": "0.0"}
-        self.barometer = {"temperature": "0.0", "pressure": "0.0", "altitude": "0.0"}
-        self.vibration = {"direction": "0", "count": "0"}
-        self.magnetometer = {"x": "0.0", "y": "0.0", "z": "0.0"}
-        self.MOG = {"Mox": "0.0", "Moy": "0.0", "Moz": "0.0"}
-        self.accelerometer = {"x": "0.0", "y": "0.0", "z": "0.0"}
-        self.AOL = {"Aox": "0.0", "Aoy": "0.0", "Aoz": "0.0"}
-        self.location = {"latitude": "0.0", "longitude": "0.0", "altitude": "0.0"}
-        self.timing = {"uptime": "0", "counter_frequency": "0", "time_string": "0"}
-        self.status = {"queue_size": "0", "missed_events": "0", "buffer_error": "0", "temp_status": "0",
-                       "baro_status": "0", "accel_status": "0", "mag_status": "0", "gps_status": "0"}
-        self.event = {"event_number": "0", "counter_frequency": "0", "ticks": "0", "timestamp": "0.0",
-                      "adc": "[[0,0,0,0,0,0,0,0],[0,0,0,0,0,0,0,0]]"}
-        self.date = {"date": "s"}
-
-        # Add ons
-
-        self.date = {"date": "s"}  # Date
-        self.sequence = {"number": "0"}  # Sequence number
-        self.PAT = {"Pat": "s", "Ntf": "0"}  # Pushover application token
-
-        # Now build the main dictionary with one entry for each json string we will process
-
-        self.recd = {"temperature": self.temperature, "barometer": self.barometer, "vibration": self.vibration,
-                     "magnetometer": self.magnetometer, "MOG": self.MOG,
-                     "accelerometer": self.accelerometer, "AOL": self.AOL, "location": self.location,
-                     "timing": self.timing, "status": self.status,
-                     "event": self.event, "date": self.date, "sequence": self.sequence, "PAT": self.PAT}
-
-        self.newpat = False
-        self.newsqn = False
-
-    # Convert the incomming json strings into entries in the dictionary
-
-    def parse(self, line):  # parse the incomming json strings from arduino
-        nstr = line.replace('\n', '')  # Throw away <crtn>, we dont want them
-        try:
-            dic = ast.literal_eval(nstr)  # Build a dictionary entry
-            kys = dic.keys()  # Get key names, the first is the address
-            if self.recd.has_key(kys[0]):  # Check we know about records with this key
-                self.recd[kys[0]] = dic[kys[0]]  # and put it in the dictionary at that address
-
-            if kys[0] == "PAT":
-                self.newpat = True
-
-            if kys[0] == "sequence":
-                self.newsqn = True
-
-        except Exception, e:
-            pass  # Didnt understand, throw it away
-
-    # Here we just return dictionaries
-
-    def get_vib(self):
-        return self.recd["vibration"]
-
-    def get_tim(self):
-        return self.recd["timing"]
-
-    def get_loc(self):
-        return self.recd["location"]
-
-    def get_sts(self):
-        return self.recd["status"]
-
-    def get_bmp(self):
-        return self.recd["barometer"]
-
-    def get_acl(self):
-        return self.recd["accelerometer"]
-
-    def get_mag(self):
-        return self.recd["magnetometer"]
-
-    def get_htu(self):
-        return self.recd["temperature"]
-
-    def get_evt(self):
-        return self.recd["event"]
-
-    def get_dat(self):
-        return self.recd["date"]
-
-    def get_sqn(self):
-        return self.recd["sequence"]
-
-    def get_pat(self):
-        return self.recd["PAT"]
+from event import Event
+from socket import Socket_io
+from keyboard import KeyBoard
+from notifications import Notifications
+from registrations import Registrations
 
 
 def main():
@@ -302,11 +71,8 @@ def main():
     kbrd.echo_off()
 
     sio = Socket_io(ipport)
-
     evt = Event()
-
     nfs = Notifications()
-
     reg = Registrations()
 
     newsqn = False
